refactor: align naming convention
