+++ src/ByteSync.ServerCommon/Interfaces/Services/Storage/IBlobStorageContainerService.cs
@@ -1,7 +1,7 @@
 ﻿using Azure.Storage;
 using Azure.Storage.Blobs;
 
-namespace ByteSync.ServerCommon.Interfaces.Services;
+namespace ByteSync.ServerCommon.Interfaces.Services.Storage;
 
 public interface IBlobStorageContainerService
 {
