+++ tests/ByteSync.ServerCommon.Tests/Commands/Storage/CleanupAzureBlobStorageSnippetsCommandHandlerTests.cs
@@ -3,7 +3,7 @@ using Azure.Storage.Blobs;
 using Azure.Storage.Blobs.Models;
 using ByteSync.ServerCommon.Business.Settings;
 using ByteSync.ServerCommon.Commands.Storage;
-using ByteSync.ServerCommon.Interfaces.Services;
+using ByteSync.ServerCommon.Interfaces.Services.Storage;
 using FakeItEasy;
 using FluentAssertions;
 using Microsoft.Extensions.Logging;
