+++ tests/ByteSync.ServerCommon.Tests/Commands/Storage/CleanupCloudflareR2SnippetsCommandHandlerTests.cs
@@ -2,7 +2,7 @@ using Amazon.S3;
 using Amazon.S3.Model;
 using ByteSync.ServerCommon.Business.Settings;
 using ByteSync.ServerCommon.Commands.Storage;
-using ByteSync.ServerCommon.Interfaces.Services;
+using ByteSync.ServerCommon.Interfaces.Services.Storage;
 using FakeItEasy;
 using FluentAssertions;
 using Microsoft.Extensions.Logging;
